refactor: separate python tilelang kernels from aot descriptors.
- initialize platform kernel packages explicitly after native operator registration
- move TileLang DSL/JIT implementations under kernels_npu and keep compiler-owned AOT descriptors thin
- preserve AOT registry behavior while updating cache dependencies, documentation, and tests
