+++ xllm/python/kernels_npu/tilelang/spec_verify_token_update.py
@@ -17,7 +17,6 @@
 import tilelang
 import tilelang.language as T
 
-from ....common.spec import DispatchField, TilelangKernel, register_kernel
 from .utils import DEFAULT_ASCEND_PASS_CONFIGS, SUPPORTED_SPEC_VERIFY_WIDTHS
 
 SYMBOL_BUFFER_CAPACITY = T.symbolic("buffer_capacity")
@@ -75,19 +74,3 @@ def build_spec_verify_token_update_kernel(spec_width: int):
                         persistent_tokens[output_offset + 5] = T.Cast("int32", draft_token_4[batch_idx])
 
     return spec_verify_token_update
-
-
-@register_kernel
-class SpecVerifyTokenUpdateKernel(TilelangKernel):
-    DISPATCH_SCHEMA = [DispatchField("spec_width", "int32")]
-    SPECIALIZATIONS = [
-        {"variant_key": f"token_w{spec_width}", "spec_width": spec_width} for spec_width in SUPPORTED_SPEC_VERIFY_WIDTHS
-    ]
-
-    @staticmethod
-    def generate_source(spec_width: int) -> str:
-        tilelang.disable_cache()
-        kernel = build_spec_verify_token_update_kernel(spec_width)
-        with tilelang.tvm.transform.PassContext(opt_level=3, config=DEFAULT_ASCEND_PASS_CONFIGS):
-            lowered = tilelang.engine.lower(kernel)
-        return lowered.kernel_source
